moves gameconfig file
