+++ src/bbc_game/configs/__init__.py
@@ -1,1 +1,2 @@
 from bbc_game.configs.default.default_game import default_game_config_factory
+from bbc_game.configs.game_config import GameConfigFactory, GameConfig
